rename conversion -> database/db2sdk
