--- coderd/conversion/doc.go
@@ -1,2 +1,0 @@
-// Package conversion provides common conversion routines from database types to codersdk types
-package conversion
